Renamed comand_check_stderr to command_check_error which will always raise an exception for a nonzero exit code and will also fail (by default) for output to stderr. This can be disabled by passing fail_on_stderr to command_check_error.
This is necessary for me because npm and other frontend build tools that we
use routinely output warnings and other non-critical, non-build-breaking
warnings and notices to stderr. We still need a way to kill the build
process, on true errors, so we must rely on exit codes.

A test for this change is in test_snap.py.